Moved EmptyFileException to exceptions.files package
