+++ src/main/java/com/blbulyandavbulyan/socialmedia/exceptions/files/EmptyFileException.java
@@ -1,5 +1,6 @@
-package com.blbulyandavbulyan.socialmedia.exceptions;
+package com.blbulyandavbulyan.socialmedia.exceptions.files;
 
+import com.blbulyandavbulyan.socialmedia.exceptions.SocialMediaException;
 import org.springframework.http.HttpStatus;
 
 public class EmptyFileException extends SocialMediaException {
