+++ src/test/java/com/blbulyandavbulyan/socialmedia/services/FileServiceTest.java
@@ -3,10 +3,9 @@ package com.blbulyandavbulyan.socialmedia.services;
 import com.blbulyandavbulyan.socialmedia.configs.FileConfigurationProperties;
 import com.blbulyandavbulyan.socialmedia.entites.File;
 import com.blbulyandavbulyan.socialmedia.entites.User;
-import com.blbulyandavbulyan.socialmedia.exceptions.EmptyFileException;
+import com.blbulyandavbulyan.socialmedia.exceptions.files.EmptyFileException;
 import com.blbulyandavbulyan.socialmedia.repositories.FileRepository;
 import com.blbulyandavbulyan.socialmedia.utils.ExtensionResolver;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
